backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow>=1.2.0
marshmallow-sqlalchemy>=1.0.0
psycopg[binary]
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
cryptography==41.0.7
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.1.0
httpx==0.26.0
gunicorn==21.2.0
tiktoken>=0.7.0
pytest>=7.0.0
pytest-flask>=1.3.0